requirements.txt
torch
numpy
scipy
matplotlib
wandb
pyyaml